refactor: split automation.rs into a module directory by concern [T20260320-063934]
automation.rs was a 1317-line flat module mixing five distinct concerns.
Split into nine focused submodules:

- mod.rs       — execute() dispatch + AutomationExecutor
- task.rs      — start_task, update_task
- worktree.rs  — create_task_worktree, finalize_task_worktree + helpers
- commit.rs    — commit_task_changes, conflict detection
- pr.rs        — open_pr_from_task, merge_pr_from_task + tests
- git.rs       — low-level git process helpers
- freshness.rs — BranchFreshness, ensure_branch_fresh_against_base
- review.rs    — fetch_review_decision_from_gh
- input.rs     — input parsing, path canonicalization

Also collapses a nested if in job_runner.rs flagged by clippy.
