+++ orbit-engine/src/executor/automation/git.rs
@@ -1,0 +1,134 @@
+use std::path::Path;
+
+use orbit_exec::{EnvironmentMode, ExecRequest, NoSandbox, StdinMode, run_process};
+use orbit_types::OrbitError;
+
+pub(super) fn git_output_paths(
+    current_dir: &Path,
+    args: &[&str],
+) -> Result<Vec<String>, OrbitError> {
+    let raw = git_output_raw(current_dir, args)?;
+    Ok(raw
+        .split('\0')
+        .filter(|value| !value.is_empty())
+        .map(ToOwned::to_owned)
+        .collect())
+}
+
+pub(super) fn git_output(current_dir: &Path, args: &[&str]) -> Result<String, OrbitError> {
+    Ok(git_output_raw(current_dir, args)?.trim().to_string())
+}
+
+pub(super) fn git_output_raw(current_dir: &Path, args: &[&str]) -> Result<String, OrbitError> {
+    let result = run_process(
+        &ExecRequest {
+            program: "git".to_string(),
+            args: args.iter().map(|value| (*value).to_string()).collect(),
+            current_dir: Some(current_dir.to_string_lossy().to_string()),
+            timeout_ms: Some(30_000),
+            stdin_mode: StdinMode::Null,
+            environment_mode: EnvironmentMode::Inherit,
+            debug: false,
+        },
+        &NoSandbox,
+    )?;
+
+    if !result.success {
+        return Err(OrbitError::Execution(format!(
+            "git {} failed in '{}': {}",
+            args.join(" "),
+            current_dir.display(),
+            result.stderr.trim()
+        )));
+    }
+
+    Ok(result.stdout)
+}
+
+pub(super) fn git_success(current_dir: &Path, args: &[&str]) -> Result<(), OrbitError> {
+    let result = run_process(
+        &ExecRequest {
+            program: "git".to_string(),
+            args: args.iter().map(|value| (*value).to_string()).collect(),
+            current_dir: Some(current_dir.to_string_lossy().to_string()),
+            timeout_ms: Some(30_000),
+            stdin_mode: StdinMode::Null,
+            environment_mode: EnvironmentMode::Inherit,
+            debug: false,
+        },
+        &NoSandbox,
+    )?;
+
+    if !result.success {
+        return Err(OrbitError::Execution(format!(
+            "git {} failed in '{}': {}",
+            args.join(" "),
+            current_dir.display(),
+            result.stderr.trim()
+        )));
+    }
+
+    Ok(())
+}
+
+pub(super) fn git_command_success(
+    current_dir: &Path,
+    args: &[&str],
+) -> Result<bool, OrbitError> {
+    let result = run_process(
+        &ExecRequest {
+            program: "git".to_string(),
+            args: args.iter().map(|value| (*value).to_string()).collect(),
+            current_dir: Some(current_dir.to_string_lossy().to_string()),
+            timeout_ms: Some(30_000),
+            stdin_mode: StdinMode::Null,
+            environment_mode: EnvironmentMode::Inherit,
+            debug: false,
+        },
+        &NoSandbox,
+    )?;
+    Ok(result.success)
+}
+
+pub(super) fn fetch_remote_base(repo_root: &Path, base: &str) {
+    let _ = run_process(
+        &ExecRequest {
+            program: "git".to_string(),
+            args: vec!["fetch".to_string(), "origin".to_string(), base.to_string()],
+            current_dir: Some(repo_root.to_string_lossy().to_string()),
+            timeout_ms: Some(60_000),
+            stdin_mode: StdinMode::Null,
+            environment_mode: EnvironmentMode::Inherit,
+            debug: false,
+        },
+        &NoSandbox,
+    );
+}
+
+pub(super) fn resolve_worktree_start_point(
+    repo_root: &Path,
+    base: &str,
+) -> Result<String, OrbitError> {
+    let remote_base = format!("origin/{base}");
+    if git_command_success(
+        repo_root,
+        &[
+            "rev-parse",
+            "--verify",
+            &format!("{remote_base}^{{commit}}"),
+        ],
+    )? {
+        return Ok(remote_base);
+    }
+
+    if git_command_success(
+        repo_root,
+        &["rev-parse", "--verify", &format!("{base}^{{commit}}")],
+    )? {
+        return Ok(base.to_string());
+    }
+
+    Err(OrbitError::Execution(format!(
+        "unable to resolve base ref '{base}' for task worktree creation"
+    )))
+}
